+++ pyproject.toml
@@ -85,17 +85,63 @@ all = [
 
 [tool.setuptools]
 package-dir = {"" = "src"}
-packages = ["mxcp", "mxcp.cli", "mxcp.config", "mxcp.config.config_schemas", "mxcp.dbt", "mxcp.drift", "mxcp.drift.drift_schemas", "mxcp.endpoints", "mxcp.endpoints.endpoint_schemas", "mxcp.evals", "mxcp.evals.eval_schemas", "mxcp.plugins", "mxcp.policy", "mxcp.runtime", "mxcp.sdk", "mxcp.sdk.audit", "mxcp.sdk.auth", "mxcp.sdk.auth.providers", "mxcp.sdk.core", "mxcp.sdk.core.analytics", "mxcp.sdk.core.config", "mxcp.sdk.core.config.resolvers", "mxcp.sdk.evals", "mxcp.sdk.executor", "mxcp.sdk.executor.plugins", "mxcp.sdk.executor.plugins.duckdb_plugin", "mxcp.sdk.executor.plugins.python_plugin", "mxcp.sdk.policy", "mxcp.sdk.validator", "mxcp.sdk.validator.schemas"]
+packages = [
+    "mxcp",
+    # Core modules
+    "mxcp.core",
+    "mxcp.core.auth",
+    "mxcp.core.config",
+    "mxcp.core.refs",
+    # Definitions
+    "mxcp.definitions",
+    "mxcp.definitions.endpoints",
+    "mxcp.definitions.evals",
+    # Executor
+    "mxcp.executor",
+    "mxcp.executor.runners",
+    "mxcp.executor.session",
+    # Interfaces
+    "mxcp.interfaces",
+    "mxcp.interfaces.cli",
+    "mxcp.interfaces.server",
+    # Other core modules
+    "mxcp.plugins",
+    "mxcp.runtime",
+    "mxcp.schemas",
+    # Services
+    "mxcp.services",
+    "mxcp.services.audit",
+    "mxcp.services.dbt",
+    "mxcp.services.drift",
+    "mxcp.services.endpoints",
+    # SDK
+    "mxcp.sdk",
+    "mxcp.sdk.audit",
+    "mxcp.sdk.audit.backends",
+    "mxcp.sdk.auth",
+    "mxcp.sdk.auth.providers",
+    "mxcp.sdk.core",
+    "mxcp.sdk.core.analytics",
+    "mxcp.sdk.core.config",
+    "mxcp.sdk.core.config.resolvers",
+    "mxcp.sdk.core.config.schemas",
+    "mxcp.sdk.evals",
+    "mxcp.sdk.executor",
+    "mxcp.sdk.executor.plugins",
+    "mxcp.sdk.executor.plugins.duckdb_plugin",
+    "mxcp.sdk.executor.plugins.python_plugin",
+    "mxcp.sdk.policy",
+    "mxcp.sdk.validator",
+    "mxcp.sdk.validator.decorators",
+    "mxcp.sdk.validator.decorators.schemas",
+]
 include-package-data = true
 
 [tool.setuptools.package-data]
 "mxcp" = ["py.typed"]
-"mxcp.config.config_schemas" = ["*.json"]
-"mxcp.endpoints.endpoint_schemas" = ["*.json"]
-"mxcp.drift.drift_schemas" = ["*.json"]
-"mxcp.evals.eval_schemas" = ["*.json"]
+"mxcp.schemas" = ["*.json"]
 "mxcp.sdk.core.config.schemas" = ["*.json"]
-"mxcp.sdk.validator.schemas" = ["*.json"]
+"mxcp.sdk.validator.decorators.schemas" = ["*.json"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
