+++ src/mxcp/sdk/validator/__init__.py
@@ -1,12 +1,10 @@
 """MXCP SDK Validator - OpenAPI-style type validation and conversion.
 
-This module provides comprehensive type validation functionality including:
+This module provides core type validation functionality used by MXCP internally:
 - Input parameter validation with default values and constraints
 - Output result validation with schema compliance
 - Type conversion between Python and OpenAPI types
 - Sensitive field masking and data protection
-- Validation decorators for functions and methods
-- Schema loading from files and validation
 
 ## Key Components
 
@@ -20,16 +18,14 @@ This module provides comprehensive type validation functionality including:
 - `ParameterSchema`: Individual parameter definition
 - `TypeSchema`: Type definitions (string, number, object, array, etc.)
 
-### Decorators
-- `@validate`: Full validation decorator with input and output schemas
-- `@validate_input`: Validate only input parameters
-- `@validate_output`: Validate only output results
-- `@validate_strict`: Strict validation with no type coercion
+## High-Level Decorators
 
-### Schema Loading
-- `load_schema`: Load schema from string content
-- `load_schema_from_file`: Load schema from YAML/JSON file
-- `validate_schema_structure`: Validate schema structure
+For decorator-based validation and schema loading utilities, see the
+`mxcp.sdk.validator.decorators` submodule:
+
+```python
+from mxcp.sdk.validator.decorators import validate, validate_input, validate_output
+```
 
 ## Quick Examples
 
@@ -91,26 +87,7 @@ validated_output = validator.validate_output(result)
 # sensitive_data is automatically masked: "[REDACTED]"
 ```
 
-### Using Decorators
-```python
-from mxcp.sdk.validator import validate
 
-# Inline schema validation
-@validate(
-    input_schema=[
-        {"name": "user_id", "type": "string", "required": True},
-        {"name": "limit", "type": "integer", "default": 10}
-    ],
-    output_schema={"type": "array", "items": {"type": "object"}}
-)
-def get_users(user_id: str, limit: int = 10) -> list[dict]:
-    return [{"id": user_id, "name": "User"}]
-
-# Schema from file
-@validate.from_file("schemas/my_function.yaml")
-def my_function(x: int) -> str:
-    return str(x)
-```
 """
 
 from ._types import (
@@ -124,8 +101,6 @@ from .converters import (
     ValidationError,
 )
 from .core import TypeValidator
-from .decorators import validate, validate_input, validate_output, validate_strict
-from .loaders import load_schema, load_schema_from_file, validate_schema_structure
 
 __all__ = [
     # Types
@@ -138,13 +113,4 @@ __all__ = [
     "ValidationError",
     # Core validator
     "TypeValidator",
-    # Decorators
-    "validate",
-    "validate_input",
-    "validate_output",
-    "validate_strict",
-    # Loaders
-    "load_schema",
-    "load_schema_from_file",
-    "validate_schema_structure",
 ]
