+++ src/mxcp/sdk/validator/decorators/schemas/validation-schema-1.json
